Fix detailed placement inside the TEMP_ANALOG domain
Before, when running detailed_placement, no cells were being placed
inside the TEMP_ANALOG voltage domain, even when using the row-deletion
algorithm from OpenFASOC.

Apparently now when a voltage domain is created, not only a Region but
also a Group is created within the database with the same name (here,
TEMP_ANALOG). If this Group is deleted, detailed_placement succeeds
and places cells inside the created domain.

Moreover, if instead of deleting the Group you assign the instances
to the Group (and not to the Region, like it was being done in
read_domain_instances.tcl), detailed_placement manages to put
all cells in the corresponding regions, even without the row-deletion
tactic.